requirements.txt
streamlit>=1.14.0
azure-identity>=1.12.0
azure-keyvault-secrets>=4.6.0
pyyaml>=6.0